+++ src/index.js
@@ -91,6 +91,11 @@
             } else if (typeof userConfig.thresholds !== 'object') {
                 throw 'Thresholds must be an Object.';
             }
+
+            if (userConfig.onClick !== undefined) {
+                this.parentPointer = userConfig.onClick.parent !== undefined;
+                this.childPointer = userConfig.onClick.child !== undefined;
+            }
         }
 
         var defaultOnClick = {
@@ -120,7 +125,7 @@
                 '#cf4799', '#c42583', '#731451', '#f3d1bf', '#c77745'
             ],  // Colors to use for blocks.
             transitionTime: userConfig.transitionTime || 1500,  // Time for a transition to start and complete.
-            truncate: userConfig.truncate || 25,  // Maximum length of a parent label before it gets truncated.
+            truncate: userConfig.truncate || 0,  // Maximum length of a parent label before it gets truncated.
             sortFunction: userConfig.sortFunction || sortJson,  // A custom sort function.
             valueKeyName: userConfig.valueKeyName // Set a custom key value in the tooltip.
         };
@@ -558,6 +563,7 @@
             .style('fill', function(obj) {
                 return (obj.parentColor !== undefined) ? _this.config.colors[obj.parentColor] : '#000000';
             })
+            .style('cursor', _this.parentPointer ? 'pointer' : 'default')
             .attr('class', 'relationshipGraph-Text')
             .attr('transform', 'translate(-6, ' + _this.config.blockSize / 1.5 + ')')
             .on('click', function(obj) {
@@ -609,7 +615,8 @@
             })
             .style('fill', function(obj) {
                 return (obj.parentColor !== undefined) ? _this.config.colors[obj.parentColor] : '#000000';
-            });
+            })
+            .style('cursor', _this.childPointer ? 'pointer' : 'default');
     };
 
     /**
@@ -637,6 +644,7 @@
             .style('fill', function(obj) {
                 return _this.config.colors[obj.color % _this.config.colors.length] || _this.config.colors[0];
             })
+            .style('cursor', _this.childPointer ? 'pointer' : 'default')
             .on('mouseover', _this.tip ? _this.tip.show : noop)
             .on('mouseout', _this.tip ? _this.tip.hide : noop)
             .on('click', function(obj) {